Linux 64bit build info
git-svn-id: https://svn.apache.org/repos/asf/openoffice/devtools@1892045 13f79535-47bb-0310-9956-ffa450edef68

+++ build-scripts/4.2.0-Dev3/unxlngix6/build_aoo64bit_on_centos7.sh
@@ -80,7 +80,9 @@ fi
 
 source ./LinuxX86-64Env.Set.sh || exit 1 
 ./bootstrap || exit 1
-\rm solenv/inc/reporevision.lst
+if [ -e solenv/inc/reporevision.lst ]; then
+	\rm solenv/inc/reporevision.lst
+fi
 cd instsetoo_native
 time perl "$SOLARENV/bin/build.pl" --all -- -P7 || exit 1
 cd util
